+++ src/main/java/domain/board/BoardReader.java
@@ -1,6 +1,6 @@
 package domain.board;
 
-import domain.Position;
+import domain.game.Position;
 import domain.piece.Piece;
 
 public interface BoardReader {
