+++ src/main/java/domain/game/Position.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.strategy.Direction;
 import java.util.HashMap;
