+++ src/main/java/domain/game/Side.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.strategy.Direction;
