+++ src/main/java/domain/piece/Piece.java
@@ -1,8 +1,8 @@
 package domain.piece;
 
-import domain.Destinations;
-import domain.Position;
-import domain.Side;
+import domain.game.Destinations;
+import domain.game.Position;
+import domain.game.Side;
 import domain.board.BoardReader;
 import domain.strategy.MovementStrategy;
 import domain.strategy.Path;
