+++ src/main/java/domain/strategy/SequenceStrategy.java
@@ -1,6 +1,6 @@
 package domain.strategy;
 
-import domain.Position;
+import domain.game.Position;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
