+++ src/test/java/domain/piece/GuardTest.java
@@ -2,9 +2,9 @@ package domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.Destinations;
-import domain.Position;
-import domain.Side;
+import domain.game.Destinations;
+import domain.game.Position;
+import domain.game.Side;
 import domain.board.Board;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
